requirements.txt
tensorflow==2.12.0
numpy
pillow
flask
